Refactor import statements to use 'otx.algo.utils.utils' for InstanceData across multiple files
