+++ src/otx/algo/common/utils/assigners/max_iou_assigner.py
@@ -18,7 +18,7 @@ from otx.algo.common.utils.assigners import BboxOverlaps2D
 from otx.algo.common.utils.structures import AssignResult
 
 if TYPE_CHECKING:
-    from otx.algo.utils.mmengine_utils import InstanceData
+    from otx.algo.utils.utils import InstanceData
 
 
 class MaxIoUAssigner:
