+++ src/otx/algo/detection/heads/class_incremental_mixin.py
@@ -14,7 +14,7 @@ from otx.algo.common.utils.utils import multi_apply
 from otx.algo.detection.utils.utils import images_to_levels
 
 if TYPE_CHECKING:
-    from otx.algo.utils.mmengine_utils import InstanceData
+    from otx.algo.utils.utils import InstanceData
 
 
 class ClassIncrementalMixin:
