+++ src/otx/algo/instance_segmentation/heads/fcn_mask_head.py
@@ -30,7 +30,7 @@ GPU_MEM_LIMIT = 1024**3  # 1 GB memory limit
 
 
 if TYPE_CHECKING:
-    from otx.algo.utils.mmengine_utils import InstanceData
+    from otx.algo.utils.utils import InstanceData
 
 
 class FCNMaskHead(BaseModule):
